Add ber tag factory to parser to allow custom tag creation.
This may help to cope with data that was designed errornously such as
GlobalPlatform public key data which has 'B0' and 'F0' tags which are
not valid. Explanation from document [1] - Table 3-8:
GlobalPlatform acknowledges as an error that values 'B0' and 'F0' in the table above are not
valid ASN.1 tag values. This mistake was introduced in previous versions of this specification.
Although this may be a problem for standard tools used to edit or display certificates and ASN.1
structures in general, it’s been decided, for the moment at least, to keep such values to preserve
backward compatibility with existing implementations already coping with such errors.

[1] https://globalplatform.org/wp-content/uploads/2019/07/GPC_2.3_A_ConfidentialCardContentMgmt_v1.2_PublicRelease-replacement.pdf